1. 각 entity data타입 수정 2. jwt.secret 추가 3. 자유 게시판 전체 조회 단위 테스트 (완) 4. 댓글 백로직 작성중(미완)

+++ src/main/java/com/sick/apeuda/repository/BoardRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
     @Query("SELECT b FROM Board b WHERE b.user.email = :userId")
     List<Board> findPostsByUserId(@Param("userId") String userId);
 }
